requirements.txt
# AI Model and API Libraries
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Overleaf Integration
python-overleaf>=0.1.0

# NLP and Text Analysis
spacy>=3.7.0
nltk>=3.8.1
pandas>=2.0.0

# Web Interface
flask>=2.3.0
werkzeug>=2.3.0
jinja2>=3.1.2
flask-wtf>=1.1.1

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
